+++ CHANGELOG.md
@@ -20,6 +20,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * `coco_align` no longer handles annotations
 * MLOps can now handle input nodes with multiple connections (variable length inputs)
 * DINO Filter now marks sites as `system_rejected` instead of removing them.
+* Depth Filter now marks sites as `system_rejected` instead of removing them.
 
 
 ## Version 0.6.8 - Target 2023-05-22